[MT-1766] Consent logging improvements (#351)
* Avoid tracking consent event when user consent status is unknown

This also means that we track nothing on reset consent.

* Only delete non audit event when clearing the queue

This way we can avoid deleting those events when we decline consent and events are queued for other reasons (like connectivity being down).
Make sure those audit events are then sent when the queue is later flushed (like on connectivity coming back on)

* Make sure that consentTrackingEventName is grant_full_consent only if there are some categories inside.

* Test clear queue not removing audit events and gdpr policy track event name