[MT-1814] DeviceData + UIKit (#356)
* Only set isBatteryMonitoringEnabled once and from the main thread

This is to avoid potential deadlocks in UIKit that expects the variable to be only set from the main thread

* Make Battery and Screen reporting configurable

To allow customers to disable them, but keep them as true by default for retrocompatibility

* Use constant UIScreen.main.fixedCoordinateSpace.bounds since we were only adding them once

The other one could've been the landscape version of the bounds if the app was launched in landscape, and be like that for the rest of the run.
Users can still determine the current resolution with a combination of orientation and resolution.

* Add tests for DeviceData to disable battery and screen reporting